Fix the wwwroot `Content` and `Script` folder path construction.
When `FolderUpdate` action was executed for MVC type of project, cta creates `wwwroot` folder if not already exists, and move `Content` an `Script` folders in it.

Currently the `Content` and `Script` folder paths are constructed as follows:
`<ProjectPath>\wwwrootContent`
`<ProjectPath>\wwwrootScript`

Instead they should be like following:
`<ProjectPath>\wwwroot\Content`
`<ProjectPath>\wwwroot\Script`

This patch includes the fix for the path construction.